+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Chg #76: Allow to use any PSR logger, `NullLogger` by default (@vjik)
 - Chg #77: Remove `ServerRequestFactory` (@vjik)
 - Chg #77: Mark `SapiEmitter` as internal (@vjik)
+- Bug #54: Fix incorrect handling response that did not close its own output buffers (@vjik)
 
 ## 2.3.0 March 10, 2024